chore: move canvas commit test helper into the test file
commitCanvasRepositoryFilesForTest lived in commit_test_helpers.go, a
regular (non _test.go) source file, so a test-only helper shipped in the
production package. Move it into commit_canvas_repository_files_test.go
where it is used, keeping it out of the production binary.

Addresses review feedback questioning why the helper exists.
